requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.30.0
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
click==8.1.7
databases==0.9.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
isort==5.13.2
magic-filter==1.0.12
multidict
mypy-extensions
packaging
pathspec
platformdirs
propcache
pydantic
pydantic_core
python-dotenv
SQLAlchemy==2.0.36
typing_extensions
yarl
